feat: VisitCard DTO 작성

+++ src/main/java/com/walktowall/backend/store/OfflineStoreRepository.java
@@ -2,5 +2,5 @@ package com.walktowall.backend.store;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OfflineStoreRepository extends JpaRepository<OfflineStore, Long> {
+public interface OfflineStoreRepository extends JpaRepository<OfflineStore, Integer> {
 }
